+++ src/control_plane/engine/gateway_container_request_options.rs
@@ -6,6 +6,7 @@ use std::path::PathBuf;
 pub(crate) struct GatewayContainerRequestOptions {
     pub(super) image: String,
     pub(super) network: String,
+    pub(super) user: String,
     pub(super) certificate_path: PathBuf,
     pub(super) private_key_path: PathBuf,
     pub(super) bootstrap_config_path: PathBuf,
@@ -17,6 +18,7 @@ impl GatewayContainerRequestOptions {
     pub(crate) const fn new(
         image: String,
         network: String,
+        user: String,
         certificate_path: PathBuf,
         private_key_path: PathBuf,
         bootstrap_config_path: PathBuf,
@@ -26,6 +28,7 @@ impl GatewayContainerRequestOptions {
         Self {
             image,
             network,
+            user,
             certificate_path,
             private_key_path,
             bootstrap_config_path,